prefer float values so fractions are not swallowed

+++ activitysim/abm/models/disaggregate_accessibility.py
@@ -50,8 +50,8 @@ class DisaggregateAccessibilityAnnotateSettings(PydanticReadable, extra="forbid"
 
 class DisaggregateAccessibilitySettings(PydanticReadable, extra="forbid"):
     suffixes: DisaggregateAccessibilitySuffixes = DisaggregateAccessibilitySuffixes()
-    ORIGIN_SAMPLE_SIZE: int | float = 0
-    DESTINATION_SAMPLE_SIZE: int | float = 0
+    ORIGIN_SAMPLE_SIZE: float | int = 0
+    DESTINATION_SAMPLE_SIZE: float | int = 0
     BASE_RANDOM_SEED: int = 0
     add_size_tables: bool = True
     zone_id_names: dict[str, str] = {"index_col": "zone_id"}
